matrixmul

+++ src/cuda/gemm_tensorcores/matrixMul/matrixMul.cu
@@ -57,8 +57,8 @@
  * Matrix multiplication (CUDA Kernel) on the device: C = A * B
  * wA is A's width and wB is B's width
  */
-template <int BLOCK_SIZE> __global__ void MatrixMulCUDA(double *C, double *A,
-    double *B, int wA,
+template <int BLOCK_SIZE> __global__ void MatrixMulCUDA(float *C, float *A,
+    float *B, int wA,
     int wB) {
   // Block index
   int bx = blockIdx.x;
@@ -87,7 +87,7 @@ template <int BLOCK_SIZE> __global__ void MatrixMulCUDA(double *C, double *A,
 
   // Csub is used to store the element of the block sub-matrix
   // that is computed by the thread
-  double Csub = 0;
+  float Csub = 0;
 
   // Loop over all the sub-matrices of A and B
   // required to compute the block sub-matrix
@@ -96,11 +96,11 @@ template <int BLOCK_SIZE> __global__ void MatrixMulCUDA(double *C, double *A,
        a += aStep, b += bStep) {
     // Declaration of the shared memory array As used to
     // store the sub-matrix of A
-    __shared__ double As[BLOCK_SIZE][BLOCK_SIZE];
+    __shared__ float As[BLOCK_SIZE][BLOCK_SIZE];
 
     // Declaration of the shared memory array Bs used to
     // store the sub-matrix of B
-    __shared__ double Bs[BLOCK_SIZE][BLOCK_SIZE];
+    __shared__ float Bs[BLOCK_SIZE][BLOCK_SIZE];
 
     // Load the matrices from device memory
     // to shared memory; each thread loads
@@ -132,7 +132,7 @@ template <int BLOCK_SIZE> __global__ void MatrixMulCUDA(double *C, double *A,
   C[c + wB * ty + tx] = Csub;
 }
 
-void ConstantInit(double *data, int size, double val) {
+void ConstantInit(float *data, int size, float val) {
   for (int i = 0; i < size; ++i) {
     data[i] = val;
   }
@@ -146,26 +146,26 @@ int MatrixMultiply(int argc, char **argv,
                    const dim3 &dimsB) {
   // Allocate host memory for matrices A and B
   unsigned int size_A = dimsA.x * dimsA.y;
-  unsigned int mem_size_A = sizeof(double) * size_A;
-  double *h_A = reinterpret_cast<double *>(malloc(mem_size_A));
+  unsigned int mem_size_A = sizeof(float) * size_A;
+  float *h_A = reinterpret_cast<float *>(malloc(mem_size_A));
   unsigned int size_B = dimsB.x * dimsB.y;
-  unsigned int mem_size_B = sizeof(double) * size_B;
-  double *h_B = reinterpret_cast<double *>(malloc(mem_size_B));
+  unsigned int mem_size_B = sizeof(float) * size_B;
+  float *h_B = reinterpret_cast<float *>(malloc(mem_size_B));
 
   // Initialize host memory
   // const float valB = 0.01f;
-  const double  valA = 2.0f;
-  const double valB = 2.0f;
+  const float  valA = 2.0f;
+  const float valB = 2.0f;
   ConstantInit(h_A, size_A, valA);
   ConstantInit(h_B, size_B, valB);
   //printf("h_A = %f\n", h_A[0]);
   // Allocate device memory
-  double *d_A, *d_B, *d_C;
+  float *d_A, *d_B, *d_C;
 
   // Allocate host matrix C
   dim3 dimsC(dimsB.x, dimsA.y, 1);
-  unsigned int mem_size_C = dimsC.x * dimsC.y * sizeof(double);
-  double *h_C = reinterpret_cast<double *>(malloc(mem_size_C));
+  unsigned int mem_size_C = dimsC.x * dimsC.y * sizeof(float);
+  float *h_C = reinterpret_cast<float *>(malloc(mem_size_C));
 
   if (h_C == NULL) {
     fprintf(stderr, "Failed to allocate host matrix C!\n");
